requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.30.6
pydantic==2.8.2
pydantic-settings==2.4.0
SQLAlchemy[asyncio]==2.0.36
asyncpg==0.29.0
aiosqlite==0.20.0
httpx==0.27.2
python-multipart==0.0.9
orjson==3.10.7
numpy==1.26.4
pandas==2.2.3
torch==2.2.2
lightgbm==4.5.0
scikit-learn==1.5.2
# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
fastapi-users==13.0.0
# Rate Limiting
slowapi==0.1.9
redis==5.2.0
# Monitoring & Logging
prometheus-client==0.21.0
structlog==24.4.0
# Database Migrations
alembic==1.14.0
# Caching
aioredis==2.0.1
# WebSocket
websockets==13.1
# Retraining pipeline
schedule==1.2.0
